管理员api

+++ services/adminServes.cjs
@@ -5,7 +5,7 @@ const {
 const md5 = require("md5")
 const Admin = require("../models/admin.cjs")
 
-//获取权限的方法，返回权限的代码
+//获取权限的方法，返回权限的代码如 0001
 exports.getRank = async (uTel) => {
     const ins = await Admin.findOne({
         where: {
@@ -13,7 +13,7 @@ exports.getRank = async (uTel) => {
         }
     })
     const uRank = ins ? ins.toJSON().uRank : "0000"
-    console.log(uRank);
+    // console.log(uRank);
     return uRank
 }
 // 判断电话是否被注册 返回 true / false
@@ -24,12 +24,20 @@ exports.getTel = async (uTel) => {
         }
     })
     const hasTel = ins ? ins.toJSON().uTel : null
+    console.log(!!hasTel);
     return hasTel
 }
-
-
 exports.addAdmin = async (obj) => {
     // console.log(this.getRank(obj.uTel));
+    const isFind = await Admin.findOne({
+        where:{
+            uTel:obj.cTel
+        }
+    })
+    console.log(isFind);
+    if(!isFind){
+        return "管理员用户未找到"
+    }
     if (await this.getRank(obj.cTel) != "0001") {
         return "权限不足"
     }
@@ -45,6 +53,9 @@ exports.deleteAdmin = async (obj) => {
     if (await this.getRank(obj.cTel) != "0001") {
         return "权限不足"
     }
+    if(!(await this.getTel(obj.uTel))){
+        return "未找到用户"
+    }
     const ins = await Admin.destroy({
         where: {
             uTel: obj.uTel
@@ -57,23 +68,50 @@ exports.updateAdmin = async (obj) => {
     if (await this.getRank(obj.cTel) != "0001") {
         return "权限不足"
     }
-    const select = await Admin.findOne({
-        where: {
-            uTel: obj.uTel
-        }
-    })
-    if (select === null) {
-        return '未找到该管理员'
+    if(!(await this.getTel(obj.uTel))){
+        return "未找到用户"
     }
-    obj.uPwd = md5(obj.uPwd)
+    // const select = await Admin.findOne({
+    //     where: {
+    //         uTel: obj.uTel
+    //     }
+    // })
+    // if (select === null) {
+    //     return '未找到该管理员'
+    // }
+    // obj.uPwd && (obj.uPwd = md5(obj.uPwd))
+    console.log(obj);
     const ins = await Admin.update(obj, {
         where: {
             uTel: obj.uTel
         }
     })
-    return ins.toJSON()
+    console.log(ins);
+    return ins
 }
-exports.getAdminById = async (uTel) => {
+exports.updateOur = async (obj) => {
+    if(!(await this.getTel(obj.uTel))){
+        return "未找到用户"
+    }
+    // const select = await Admin.findOne({
+    //     where: {
+    //         uTel: obj.uTel
+    //     }
+    // })
+    // if (select === null) {
+    //     return '未找到该管理员'
+    // }
+    // obj.uPwd && (obj.uPwd = md5(obj.uPwd))
+    console.log(obj);
+    const ins = await Admin.update(obj, {
+        where: {
+            uTel: obj.uTel
+        }
+    })
+    console.log(ins);
+    return ins
+}
+exports.getAdminByUTel = async (uTel) => {
     const ins = await Admin.findOne({
         where: {
             uTel
@@ -102,9 +140,10 @@ exports.login = async (obj) => {
     ins = ins.toJSON()
     return ins
 }
-
 exports.signIn = async (obj) => {
-    if (this.getTel(obj.uTel)) {
+    console.log(obj);
+    const useTel = await this.getTel(obj.uTel)
+    if (useTel) {
         return "手机号已被注册"
     }
     let ins = await Admin.findOne({
